+++ src/Custom/Responses/Tools/ResponseTool.cs
@@ -127,7 +127,7 @@ public partial class ResponseTool
     /// <summary>
     /// Creates a new instance of the <see cref="ImageGenerationTool"/> class.
     /// </summary>
-    public static ImageGenerationTool CreateImageGenerationTool(string model, ImageGenerationToolQuality? quality = null, ImageGenerationToolSize? size = null, ImageGenerationToolOutputFileFormat? outputFileFormat = null, int? outputCompressionFactor = null, ImageGenerationToolModerationLevel? moderationLevel = null, ImageGenerationToolBackground? background = null, ImageGenerationToolInputFidelity? inputFidelity = null, ImageGenerationToolInputImageMask inputImageMask = null, int? partialImageCount = null)
+    public static ImageGenerationTool CreateImageGenerationTool(string model, ImageGenerationToolQuality? quality = null, ImageGenerationToolSize? size = null, ImageGenerationToolOutputFileFormat? outputFileFormat = null, int? outputCompressionFactor = null, ImageGenerationToolModerationLevel? moderationLevel = null, ImageGenerationToolBackground? background = null, ImageGenerationToolInputFidelity? inputFidelity = null, ImageGenerationToolInputImageMask inputImageMask = null, int? partialImageCount = null, ImageGenerationToolAction? action = null)
     {
         return new ImageGenerationTool(
             kind: InternalToolType.ImageGeneration,
@@ -141,6 +141,7 @@ public partial class ResponseTool
             background: background,
             inputFidelity: inputFidelity,
             inputImageMask: inputImageMask,
-            partialImageCount: partialImageCount);
+            partialImageCount: partialImageCount,
+            action: action);
     }
 }